chore: improve UX & tests (#4)
- add basic CI
- add basic integration tests
- minor UX and inline docs tweaks

+++ src/rock.py
@@ -4,45 +4,41 @@ import logging
 import os
 import re
 import secrets
-import tempfile
 
 import charm
 from models import Artifact, CharmSpec, Versions
-from util import DOCKER, TMP_PREFIX, clone_repo, exec, load_known_versions
+from util import DOCKER, exec, load_known_versions
 
 logger = logging.getLogger(__name__)
 
 KNOWN = load_known_versions()
 
 
-def resolve_k8s_charm_single(spec: CharmSpec, charm_dir: str | None = None) -> str:
+def resolve_k8s_charm_single(spec: CharmSpec, rev: int | str) -> tuple[str, str]:
     """Return the workload version of a given charm spec - single revision."""
-    tmp_path = tempfile.mkdtemp(dir=".", prefix=TMP_PREFIX)
-    repo = spec.repo
     image_path = spec.yaml_path
     cmd = spec.cmd
     regex = spec.regex
 
-    if not charm_dir:
-        charm_dir = clone_repo(repo, tmp_path)
+    charm_dir = charm.unpack(spec.name, rev)
 
     image = exec(f"cat {charm_dir}/metadata.yaml | yq -r '{image_path}'").strip()
     artifact = Artifact(type="rock", name=spec.name, rev=image)
     if artifact in KNOWN:
-        return KNOWN[artifact]
+        return image, KNOWN[artifact]
 
     container = f"testrock_{secrets.token_hex(8)}"
     os.system(f"{DOCKER} run -d --name {container} {image}")
     raw = exec(f"{DOCKER} exec {container} {cmd}")
 
     matches = [raw] if not regex else re.findall(regex, raw)
-    assert matches, "Can't determine workload version!"
+    if not matches:
+        return image, "unknown"
 
     # clean up
-    os.system(f"rm -rf {tmp_path}")
     os.system(f"{DOCKER} rm --force {container}")
 
-    return matches[0].strip()
+    return image, matches[0].strip()
 
 
 def resolve_k8s_charm_all(spec: CharmSpec) -> list[Versions]:
@@ -53,13 +49,11 @@ def resolve_k8s_charm_all(spec: CharmSpec) -> list[Versions]:
 
     rev_to_workload = {}
     for rev in charm_revs:
-        charm_dir = charm.unpack(_charm, rev)
-        image = exec(f"cat {charm_dir}/metadata.yaml | yq -r '{spec.yaml_path}'").strip()
         try:
-            rev_to_workload[rev] = (image, resolve_k8s_charm_single(spec, charm_dir))
+            rev_to_workload[rev] = resolve_k8s_charm_single(spec, rev)
         except Exception as e:
             logger.error(e)
-            rev_to_workload[rev] = (image, "unknown")
+            rev_to_workload[rev] = ("unknown", "unknown")
 
     versions = []
     for rel, charm_rev in charm_info.items():
